added tests, comments, minor changes

+++ task_DEV1_4/task_DEV1_4/IFlyable.cs
@@ -3,6 +3,6 @@
     interface IFlyable
     {
         void FlyTo(Coordinate newCoordinate);
-        float GetFlyTime(Coordinate CurrentCoordinate, Coordinate newCoordinate);
+        int GetFlyTime(Coordinate CurrentCoordinate, Coordinate newCoordinate);
     }
 }
